@@ -3,112 +3,39 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "export NODE_OPTIONS=--openssl-legacy-provider && vue-cli-service serve",
-    "ci": "export NODE_OPTIONS=--openssl-legacy-provider && vue-cli-service serve --mode ci",
-    "build": "export NODE_OPTIONS=--openssl-legacy-provider && vue-cli-service build",
-    "lint": "export NODE_OPTIONS=--openssl-legacy-provider && vue-cli-service lint",
-    "test:unit": "export NODE_OPTIONS=--openssl-legacy-provider && vue-cli-service test:unit"
+    "dev": "vite",
+    "test-unit": "vitest run",
+    "build": "vite build",
+    "preview": "vite preview",
+    "cypress-run": "npx cypress run",
+    "lint": "eslint . --ignore-path .gitignore"
   },
   "dependencies": {
     "@amcharts/amcharts4": "^4.10.36",
+    "@mdi/font": "7.0.96",
     "ajv": "^6.12.6",
     "axios": "1.6.2",
-    "core-js": "^2.6.12",
+    "core-js": "^3.29.0",
     "deep-equal": "^2.2.1",
     "jwt-decode": "^3.1.2",
-    "vue": "2.7.14",
-    "vue-router": "^3.0.7",
-    "vuetify": "2.6.15",
-    "vuex": "^3.5.1"
+    "roboto-fontface": "*",
+    "vue": "^3.2.0",
+    "vue-router": "^4.0.0",
+    "vuetify": "^3.0.0",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.12.1",
-    "@vue/cli-plugin-eslint": "^3.12.1",
-    "@vue/cli-plugin-unit-jest": "^3.12.1",
-    "@vue/cli-service": "^3.12.1",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "^1.3.5",
-    "@webpack-cli/info": "^0.1.6",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^23.6.0",
-    "cypress": "12.13.0",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "jest": "^24.7.1",
-    "jest-junit": "^6.3.0",
-    "obsolete-webpack-plugin": "^0.5.6",
-    "sass": "1.23.0",
-    "sass-loader": "^7.1.0",
-    "vue-cli-plugin-vuetify": "0.6.3",
-    "vue-template-compiler": "^2.7.14",
-    "vuetify-loader": "^1.9.2",
-    "webpack-cli": "^3.3.6"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "@vue/standard"
-    ],
-    "rules": {
-      "indent": [
-        "error",
-        4
-      ],
-      "space-before-function-paren": [
-        "error",
-        {
-          "anonymous": "never",
-          "named": "never",
-          "asyncArrow": "always"
-        }
-      ]
-    },
-    "parserOptions": {
-      "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "defaults",
-    "last 6 versions",
-    "not ie <= 11"
-  ],
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "jsx",
-      "json",
-      "mjs",
-      "vue"
-    ],
-    "transform": {
-      "^.+\\.vue$": "vue-jest",
-      ".+\\.(css|styl|less|sass|scss|svg|png|jpg|ttf|woff|woff2)$": "jest-transform-stub",
-      "^.+\\.(mjs|jsx?)$": "babel-jest"
-    },
-    "moduleNameMapper": {
-      "^@/(.*)$": "<rootDir>/src/$1",
-      "axios": "axios/dist/node/axios.cjs"
-    },
-    "snapshotSerializers": [
-      "jest-serializer-vue"
-    ],
-    "testMatch": [
-      "**/tests/unit/**/*.spec.(js|jsx|ts|tsx)|**/__tests__/*.(js|jsx|ts|tsx)"
-    ],
-    "testURL": "http://localhost/",
-    "reporters": [
-      "default",
-      "jest-junit"
-    ]
+    "@vitejs/plugin-vue": "^4.0.0",
+    "cypress": "13.6.2",
+    "eslint": "^8.37.0",
+    "eslint-plugin-cypress": "^2.15.1",
+    "eslint-plugin-vue": "^9.3.0",
+    "jsdom": "^23.0.1",
+    "lodash": "^4.17.21",
+    "sass": "^1.60.0",
+    "unplugin-fonts": "^1.0.3",
+    "vite": "^4.2.0",
+    "vite-plugin-vuetify": "^1.0.0",
+    "vitest": "^1.1.0"
   }
 }
