@@ -1,14 +1,14 @@
-import Vue from 'vue'
+// Plugins
+import { registerPlugins } from '@/plugins'
+
+// Components
 import App from './App.vue'
-import Axios from 'axios'
-import vuetify from './plugins/vuetify'
-import store from './store'
 
-Vue.config.productionTip = false
+// Composables
+import { createApp } from 'vue'
 
-new Vue({
-    created: function() { Axios.defaults.baseURL = process.env.VUE_APP_BASE_URL },
-    vuetify,
-    store,
-    render: h => h(App)
-}).$mount('#app')
+const app = createApp(App)
+
+registerPlugins(app)
+app.mount('#app')
+
