@@ -3,16 +3,16 @@ describe('Help page', function() {
     it('Go to Meteor', function() {
         cy.visit('/')
         cy.get('i#btnMeteor').click()
-        cy.url().should('include', '/#/motorDesign')
+        cy.url().should('include', '/motorDesign')
     })
 
     it('Should display help content', function() {
-        cy.get('i#btnHelp').click()
+        cy.get('button#btnHelp').click()
         cy.contains('Meteor Manual')
     })
 
     it('Should close help', function() {
-        cy.get('i#btnCloseHelp').click()
+        cy.get('#btnCloseHelp').click()
         cy.should('not.contain', 'Meteor Manual')
     })
 
