@@ -4,7 +4,7 @@ describe('Testing demo page', function() {
         cy.visit('/')
         cy.contains('View demo').click()
 
-        cy.url().should('include', '/#/demo')
+        cy.url().should('include', '/demo')
 
         cy.get('#throatDiameter').parent().contains('mm')
 
@@ -12,9 +12,9 @@ describe('Testing demo page', function() {
     })
 
     it('Should restore demo units', function() {
-        cy.visit('/#/motorDesign')
+        cy.visit('/motorDesign')
 
-        cy.url().should('include', '/#/motorDesign')
+        cy.url().should('include', '/motorDesign')
 
         cy.contains('IMPERIAL').click()
         cy.get('#throatDiameter').parent().contains('inch')
@@ -22,7 +22,7 @@ describe('Testing demo page', function() {
         cy.visit('/')
         cy.contains('View demo').click()
 
-        cy.url().should('include', '/#/demo')
+        cy.url().should('include', '/demo')
         cy.get('#throatDiameter').parent().contains('mm')
     })
 })
