@@ -1,261 +1,260 @@
-describe.skip('RASP export', function() {
+describe('RASP export', function() {
     it('Should export to RASP in METRIC', function() {
         cy.visit('/')
         cy.contains('View demo').click()
 
-        cy.url().should('include', '/#/demo')
-
-        cy.server()
-
-        // This is the post call we are interested in capturing
-        cy.route('POST', '/export/rasp').as('postExportRASP')
+        cy.url().should('include', '/demo')
 
         cy.get('button#btnShowRASPExport').click()
 
-        cy.get('input#motorDiameter').clear().type(80)
+        cy.get('input#motorDiameter').clear()
+        cy.get('input#motorDiameter').type(80)
+        cy.get('input#motorDiameter')
             .parent()
             .contains('mm')
-        cy.get('input#motorLength').clear().type(500)
+
+        cy.get('input#motorLength').clear()
+        cy.get('input#motorLength').type(500)
+        cy.get('input#motorLength')
             .parent()
             .contains('mm')
-        cy.get('input#motorWeight').clear().type(4.23)
+
+        cy.get('input#motorWeight').clear()
+        cy.get('input#motorWeight').type(4.23)
+        cy.get('input#motorWeight')
             .parent()
             .contains('Kg')
-        cy.get('input#delay').clear().type('0-1-P')
+
+        cy.get('input#delay').clear()
+        cy.get('input#delay').type('0-1-P')
+        cy.get('input#delay')
             .parent()
             .contains('s')
 
         cy.get('button#btnExportRASP').click()
 
-        cy.wait('@postExportRASP')
-
-        // Assert on XHR
-        cy.get('@postExportRASP').then(function(xhr) {
-            console.log(xhr)
-            expect(xhr.status).to.eq(200)
-            expect(xhr.method).to.eq('POST')
-            expect(xhr.responseBody).to.eq(`L1672 80.0 500.0 0-1-P 2.812 4.230 METEOR
+        cy.readFile('cypress/downloads/meteor-RASP.eng')
+            .should('eq', `L1607 80.0 500.0 0-1-P 2.812 4.230 METEOR
     0.0 0.0
-    0.0253 917.1436
-    0.0394 1045.1818
-    0.0527 1107.0875
-    0.0656 1142.234
-    0.0783 1165.8222
-    0.0908 1184.2304
-    0.1032 1200.2518
-    0.1156 1215.144
-    0.1278 1229.4843
-    0.14 1243.542
-    0.1521 1257.4437
-    0.1641 1271.2497
-    0.176 1284.9878
-    0.1878 1298.6708
-    0.1996 1312.3036
-    0.2113 1325.8875
-    0.2229 1339.4218
-    0.2344 1352.905
-    0.2459 1366.3348
-    0.2573 1379.709
-    0.2686 1393.0251
-    0.2799 1406.2805
-    0.2911 1419.4727
-    0.3023 1432.5991
-    0.3133 1445.6571
-    0.3244 1458.6442
-    0.3353 1471.5578
-    0.3462 1484.3954
-    0.3571 1497.1545
-    0.3679 1509.8326
-    0.3786 1522.4272
-    0.3893 1534.936
-    0.3999 1547.3564
-    0.4105 1559.6861
-    0.421 1571.9226
-    0.4315 1584.0637
-    0.4419 1596.1069
-    0.4522 1608.0499
-    0.4626 1619.8905
-    0.4729 1631.6264
-    0.4831 1643.2553
-    0.4933 1654.7749
-    0.5034 1666.1832
-    0.5135 1677.4779
-    0.5236 1688.6568
-    0.5336 1699.7178
-    0.5436 1710.6589
-    0.5535 1721.4779
-    0.5634 1732.1728
-    0.5733 1742.7415
-    0.5831 1753.182
-    0.5929 1763.4924
-    0.6027 1773.6708
-    0.6124 1783.7151
-    0.6221 1793.6235
-    0.6318 1803.3941
-    0.6414 1813.0251
-    0.651 1822.5146
-    0.6606 1831.8609
-    0.6701 1841.0622
-    0.6796 1850.1167
-    0.6891 1859.0228
-    0.6985 1867.7788
-    0.7079 1876.3829
-    0.7173 1884.8337
-    0.7267 1893.1295
-    0.7361 1901.2687
-    0.7454 1909.2498
-    0.7547 1917.0713
-    0.7639 1924.7317
-    0.7732 1932.2296
-    0.7824 1939.5635
-    0.7916 1946.732
-    0.8008 1953.7338
-    0.81 1960.5676
-    0.8191 1967.232
-    0.8283 1973.7257
-    0.8374 1980.0476
-    0.8465 1986.1964
-    0.8555 1992.171
-    0.8646 1997.9701
-    0.8736 2003.5926
-    0.8827 2009.0376
-    0.8917 2014.3038
-    0.9007 2019.3903
-    0.9096 2033.3918
-    0.9185 2042.1488
-    0.9274 2046.534
-    0.9363 2049.0317
-    0.9452 2050.6907
-    0.9541 2051.9562
-    0.963 2053.0185
-    0.9719 2053.9594
-    0.9808 2054.8142
-    0.9898 2055.5984
-    0.9987 2056.3186
-    1.0076 2056.9779
-    1.0165 2057.5775
-    1.0254 2058.1181
-    1.0343 2058.5999
-    1.0432 2059.0231
-    1.0521 2059.3877
-    1.0611 2059.6937
-    1.07 2059.9411
-    1.0789 2060.13
-    1.0878 2060.2604
-    1.0967 2060.3322
-    1.1056 2060.3455
-    1.1145 2060.3002
-    1.1234 2060.1964
-    1.1324 2060.0341
-    1.1413 2059.8133
-    1.1502 2059.5339
-    1.1591 2059.196
-    1.168 2058.7995
-    1.1769 2058.3445
-    1.1858 2057.831
-    1.1948 2057.2589
-    1.2037 2056.6283
-    1.2126 2055.9391
-    1.2215 2055.1914
-    1.2304 2054.3851
-    1.2393 2053.5202
-    1.2482 2052.5968
-    1.2571 2051.6148
-    1.266 2050.5742
-    1.2749 2049.4751
-    1.2839 2048.3173
-    1.2928 2047.101
-    1.3017 2045.826
-    1.3106 2044.4924
-    1.3195 2043.1002
-    1.3284 2041.6494
-    1.3373 2040.1399
-    1.3462 2038.5718
-    1.3551 2036.945
-    1.364 2035.2596
-    1.3729 2033.5154
-    1.3818 2031.7126
-    1.3907 2029.851
-    1.3996 2027.9307
-    1.4085 2025.9517
-    1.4174 2023.914
-    1.4263 2021.8174
-    1.4351 2019.6622
-    1.4441 2006.4666
-    1.4532 1994.0989
-    1.4623 1982.4057
-    1.4714 1971.2574
-    1.4805 1960.5444
-    1.4897 1950.1745
-    1.4989 1940.0698
-    1.5081 1930.1645
-    1.5173 1920.4031
-    1.5266 1910.7392
-    1.5359 1901.1334
-    1.5453 1891.5528
-    1.5546 1881.9696
-    1.564 1872.3605
-    1.5735 1862.7062
-    1.5829 1852.9904
-    1.5924 1843.1996
-    1.6019 1833.3227
-    1.6115 1823.3503
-    1.6211 1813.2751
-    1.6307 1803.0908
-    1.6403 1792.7925
-    1.65 1782.3763
-    1.6598 1771.8393
-    1.6695 1761.1792
-    1.6793 1750.3944
-    1.6891 1739.4839
-    1.699 1728.447
-    1.7089 1717.2837
-    1.7189 1705.9941
-    1.7289 1694.5787
-    1.7389 1683.0384
-    1.749 1671.3741
-    1.7591 1659.587
-    1.7692 1647.6785
-    1.7794 1635.6502
-    1.7897 1623.5037
-    1.8 1611.2408
-    1.8103 1598.8635
-    1.8207 1586.3738
-    1.8312 1573.7738
-    1.8417 1561.0657
-    1.8523 1548.2517
-    1.8629 1535.3341
-    1.8735 1522.3155
-    1.8842 1509.1981
-    1.895 1495.9846
-    1.9059 1482.6774
-    1.9168 1469.2793
-    1.9277 1455.7928
-    1.9387 1442.2206
-    1.9498 1428.5656
-    1.961 1414.8304
-    1.9722 1401.0178
-    1.9835 1387.1309
-    1.9948 1373.1723
-    2.0062 1359.1451
-    2.0177 1345.0522
-    2.0293 1330.8967
-    2.0409 1316.6814
-    2.0527 1302.4096
-    2.0645 1288.0841
-    2.0763 1273.7083
-    2.0854 1098.7741
-    2.0915 825.4024
-    2.0977 618.2182
-    2.1038 461.4383
-    2.1099 343.0003
-    2.1161 253.6921
-    2.1222 186.4836
-    2.1284 136.013
-    2.1345 98.193
-    2.1406 69.9073
-    2.1468 48.7756
-    2.1529 32.9662`)
-        })
+    0.0253 889.1291
+    0.0394 1011.0918
+    0.0527 1069.9741
+    0.0656 1103.3847
+    0.0783 1125.8047
+    0.0908 1143.3014
+    0.1033 1158.5299
+    0.1156 1172.6846
+    0.1279 1186.3136
+    0.1401 1199.6723
+    0.1521 1212.8812
+    0.1642 1225.997
+    0.1761 1239.0465
+    0.1879 1252.0415
+    0.1997 1264.9868
+    0.2114 1277.8837
+    0.223 1290.7317
+    0.2346 1303.529
+    0.2461 1316.274
+    0.2575 1328.9642
+    0.2688 1341.5975
+    0.2801 1354.1714
+    0.2913 1366.6836
+    0.3025 1379.1316
+    0.3135 1391.5131
+    0.3246 1403.8256
+    0.3355 1416.0669
+    0.3465 1428.2346
+    0.3573 1440.3263
+    0.3681 1452.3397
+    0.3788 1464.2725
+    0.3895 1476.1225
+    0.4002 1487.8873
+    0.4107 1499.5649
+    0.4213 1511.1528
+    0.4317 1522.649
+    0.4422 1534.0512
+    0.4525 1545.3573
+    0.4629 1556.5652
+    0.4732 1567.6727
+    0.4834 1578.6779
+    0.4936 1589.5785
+    0.5038 1600.4502
+    0.5139 1611.3051
+    0.5239 1622.0429
+    0.534 1632.6621
+    0.544 1643.1606
+    0.5539 1653.5368
+    0.5638 1663.7889
+    0.5737 1673.9151
+    0.5835 1683.9138
+    0.5933 1693.7832
+    0.6031 1703.5218
+    0.6128 1713.1278
+    0.6225 1722.5996
+    0.6322 1731.9357
+    0.6418 1741.1345
+    0.6514 1750.1945
+    0.661 1759.114
+    0.6706 1767.8917
+    0.6801 1776.5261
+    0.6895 1785.0157
+    0.699 1793.3591
+    0.7084 1801.5548
+    0.7178 1809.6016
+    0.7272 1817.4981
+    0.7366 1825.2429
+    0.7459 1832.8347
+    0.7552 1840.2723
+    0.7645 1847.5544
+    0.7737 1854.6799
+    0.783 1861.6474
+    0.7922 1868.4559
+    0.8014 1875.1041
+    0.8105 1881.5909
+    0.8197 1887.9153
+    0.8288 1894.0762
+    0.838 1900.0725
+    0.847 1905.9032
+    0.8561 1911.5673
+    0.8652 1917.0638
+    0.8742 1922.3918
+    0.8833 1927.5504
+    0.8923 1932.5386
+    0.9013 1937.3556
+    0.9103 1945.6677
+    0.9192 1957.7061
+    0.9281 1963.4556
+    0.937 1966.5057
+    0.9459 1968.372
+    0.9548 1969.6985
+    0.9637 1970.7602
+    0.9726 1971.6755
+    0.9815 1972.4955
+    0.9904 1973.2425
+    0.9993 1973.9264
+    1.0082 1974.5513
+    1.0172 1975.1193
+    1.0261 1975.6311
+    1.035 1976.0871
+    1.0439 1976.4877
+    1.0528 1976.8327
+    1.0617 1977.1223
+    1.0706 1977.3565
+    1.0795 1977.5353
+    1.0885 1977.6587
+    1.0974 1977.7266
+    1.1063 1977.7393
+    1.1152 1977.6965
+    1.1241 1977.5983
+    1.133 1977.4447
+    1.1419 1977.2358
+    1.1509 1976.9714
+    1.1598 1976.6517
+    1.1687 1976.2766
+    1.1776 1975.846
+    1.1865 1975.3601
+    1.1954 1974.8187
+    1.2043 1974.222
+    1.2132 1973.5698
+    1.2221 1972.8621
+    1.2311 1972.0991
+    1.24 1971.2805
+    1.2489 1970.4066
+    1.2578 1969.4771
+    1.2667 1968.4921
+    1.2756 1967.4517
+    1.2845 1966.3557
+    1.2934 1965.2043
+    1.3023 1963.9972
+    1.3112 1962.7347
+    1.3201 1961.4165
+    1.329 1960.0428
+    1.3379 1958.6134
+    1.3468 1957.1285
+    1.3557 1955.5879
+    1.3646 1953.9916
+    1.3735 1952.3396
+    1.3824 1950.6319
+    1.3913 1948.8685
+    1.4002 1947.0493
+    1.4091 1945.1744
+    1.418 1943.2436
+    1.4269 1941.257
+    1.4358 1936.5086
+    1.4448 1924.1685
+    1.4539 1912.5834
+    1.463 1901.6122
+    1.4721 1891.1359
+    1.4812 1881.0536
+    1.4904 1871.2804
+    1.4996 1861.7444
+    1.5088 1852.385
+    1.5181 1843.1512
+    1.5274 1833.9997
+    1.5367 1824.8945
+    1.546 1815.8048
+    1.5554 1806.705
+    1.5648 1797.5736
+    1.5743 1788.3925
+    1.5837 1779.1465
+    1.5932 1769.8231
+    1.6027 1760.4118
+    1.6123 1750.9039
+    1.6219 1741.2924
+    1.6315 1731.5713
+    1.6412 1721.736
+    1.6509 1711.7828
+    1.6606 1701.7086
+    1.6704 1691.5114
+    1.6802 1681.1893
+    1.69 1670.7413
+    1.6999 1660.1667
+    1.7098 1649.4652
+    1.7198 1638.6368
+    1.7298 1627.6817
+    1.7398 1616.6005
+    1.7499 1605.394
+    1.76 1594.0872
+    1.7702 1582.8196
+    1.7804 1571.4372
+    1.7907 1559.9417
+    1.801 1548.3348
+    1.8113 1536.618
+    1.8217 1524.7935
+    1.8322 1512.863
+    1.8427 1500.8287
+    1.8533 1488.6927
+    1.8639 1476.457
+    1.8745 1464.124
+    1.8853 1451.6959
+    1.8961 1439.175
+    1.9069 1426.5637
+    1.9178 1413.8645
+    1.9288 1401.0797
+    1.9398 1388.212
+    1.9509 1375.2639
+    1.962 1362.2379
+    1.9733 1349.1367
+    1.9846 1335.9629
+    1.9959 1322.7192
+    2.0073 1309.4084
+    2.0188 1296.0332
+    2.0304 1282.5964
+    2.0421 1269.1008
+    2.0538 1255.5492
+    2.0656 1241.9446
+    2.0775 1228.2898
+    2.0866 1061.9828
+    2.0927 801.2491
+    2.0988 602.7675
+    2.105 451.8989
+    2.1111 337.4081
+    2.1173 250.6774
+    2.1234 185.1015
+    2.1295 135.6203
+    2.1357 98.3591
+    2.1418 70.3497
+    2.148 49.3132
+    2.1541 33.4867`)
     })
 })
