@@ -56,6 +56,9 @@ function applyCommonCheck(formValues, units) {
         formValues.name = `Motor-${Number(Math.random() * 1000).toFixed(0)}`
     }
     cy.contains(units).click()
+    cy.get('#throatDiameter').click()
+    cy.get('#throatDiameter').parent().contains(getLengthUnit(units))
+
     cy.get('input#name').should('have.value', '')
 
     cy.get('input#name').type(formValues.name)
@@ -77,7 +80,9 @@ Cypress.Commands.add('runSavedMotor', (motorName) => {
 })
 
 Cypress.Commands.add('saveAndRunCheck', (motorName, expectedResults) => {
-    cy.get('input#name').clear().type(motorName)
+    cy.get('input#name').clear()
+    cy.get('input#name').type(motorName)
+
     cy.saveMotor()
     cy.contains('Reset').click()
     cy.runSavedMotor(motorName)
@@ -87,46 +92,46 @@ Cypress.Commands.add('saveAndRunCheck', (motorName, expectedResults) => {
 Cypress.Commands.add('fillForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
-    cy.get('div.v-list-item__title').contains(formValues.propellantId).click()
+    cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
-        .contains('Hollow cylinder').click()
+    cy.contains('Hollow cylinder').click()
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(2).click()
-    cy.get('div.menuable__content__active')
-        .contains(formValues.coreSurface).click()
+    cy.get('.v-overlay__content > .v-list').contains(formValues.coreSurface).click()
 
     cy.get('@coreSurfaces').eq(3).click()
-    cy.get('div.menuable__content__active')
-        .contains(formValues.outerSurface).click()
+    cy.get('.v-overlay__content > .v-list').contains(formValues.outerSurface).click()
 
     cy.get('@coreSurfaces').eq(4).click()
-    cy.get('div.menuable__content__active')
-        .contains(formValues.endsSurface).click()
+    cy.get('.v-overlay__content > .v-list').contains(formValues.endsSurface).click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#coreDiameter')
         .type(formValues.coreDiameter)
+    cy.get('input#coreDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#outerDiameter')
         .type(formValues.outerDiameter)
+    cy.get('input#outerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#segmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#segmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -135,11 +140,13 @@ Cypress.Commands.add('fillForm', (formValues, units, submit = true) => {
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -151,39 +158,43 @@ Cypress.Commands.add('fillForm', (formValues, units, submit = true) => {
 Cypress.Commands.add('fillFinocylForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
     cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains('Finocyl').click()
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces')
         .eq(2).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains(formValues.endsSurface).click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#finocylSegmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#finocylSegmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -192,21 +203,25 @@ Cypress.Commands.add('fillFinocylForm', (formValues, units, submit = true) => {
 
     cy.get('input#finocylOuterDiameter')
         .type(formValues.outerDiameter)
+    cy.get('input#finocylOuterDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#finocylInnerDiameter')
         .type(formValues.innerDiameter)
+    cy.get('input#finocylInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#finocylFinWidth')
         .type(formValues.finWidth)
+    cy.get('input#finocylFinWidth')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#finocylFinDiameter')
         .type(formValues.finDiameter)
+    cy.get('input#finocylFinDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -221,39 +236,43 @@ Cypress.Commands.add('fillFinocylForm', (formValues, units, submit = true) => {
 Cypress.Commands.add('fillStarForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
     cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains('Star').click()
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces')
         .eq(2).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains(formValues.endsSurface).click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#starSegmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#starSegmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -262,16 +281,19 @@ Cypress.Commands.add('fillStarForm', (formValues, units, submit = true) => {
 
     cy.get('input#starOuterDiameter')
         .type(formValues.outerDiameter)
+    cy.get('input#starOuterDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#starInnerDiameter')
         .type(formValues.innerDiameter)
+    cy.get('input#starInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#starPointDiameter')
         .type(formValues.pointDiameter)
+    cy.get('input#starPointDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -286,39 +308,43 @@ Cypress.Commands.add('fillStarForm', (formValues, units, submit = true) => {
 Cypress.Commands.add('fillMoonBurnerForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
     cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains('Moon burner').click()
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces')
         .eq(2).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains(formValues.endsSurface).click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#moonBurnerSegmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#moonBurnerSegmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -327,16 +353,19 @@ Cypress.Commands.add('fillMoonBurnerForm', (formValues, units, submit = true) =>
 
     cy.get('input#moonBurnerOuterDiameter')
         .type(formValues.outerDiameter)
+    cy.get('input#moonBurnerOuterDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#moonBurnerCoreDiameter')
         .type(formValues.coreDiameter)
+    cy.get('input#moonBurnerCoreDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#moonBurnerCoreOffset')
         .type(formValues.coreOffset)
+    cy.get('input#moonBurnerCoreOffset')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -348,39 +377,43 @@ Cypress.Commands.add('fillMoonBurnerForm', (formValues, units, submit = true) =>
 Cypress.Commands.add('fillCSlotForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
     cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains('C slot').click()
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces')
         .eq(2).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains(formValues.endsSurface).click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#cSlotSegmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#cSlotSegmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -389,26 +422,31 @@ Cypress.Commands.add('fillCSlotForm', (formValues, units, submit = true) => {
 
     cy.get('input#cSlotOuterDiameter')
         .type(formValues.outerDiameter)
+    cy.get('input#cSlotOuterDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#cSlotCoreDiameter')
         .type(formValues.coreDiameter)
+    cy.get('input#cSlotCoreDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#cSlotSlotWidth')
         .type(formValues.slotWidth)
+    cy.get('input#cSlotSlotWidth')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#cSlotSlotDepth')
         .type(formValues.slotDepth)
+    cy.get('input#cSlotSlotDepth')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#cSlotSlotOffset')
         .type(formValues.slotOffset)
+    cy.get('input#cSlotSlotOffset')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -420,39 +458,43 @@ Cypress.Commands.add('fillCSlotForm', (formValues, units, submit = true) => {
 Cypress.Commands.add('fillRodTubeForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
     cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains('Rod and tube').click()
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces')
         .eq(2).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains(formValues.endsSurface).click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#rodTubeSegmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#rodTubeSegmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -461,16 +503,19 @@ Cypress.Commands.add('fillRodTubeForm', (formValues, units, submit = true) => {
 
     cy.get('input#rodTubeRodDiameter')
         .type(formValues.rodDiameter)
+    cy.get('input#rodTubeRodDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#rodTubeTubeOuterDiameter')
         .type(formValues.tubeOuterDiameter)
+    cy.get('input#rodTubeTubeOuterDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#rodTubeTubeInnerDiameter')
         .type(formValues.tubeInnerDiameter)
+    cy.get('input#rodTubeTubeInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -489,57 +534,67 @@ Cypress.Commands.add('setMotorSimAdvancedConfig', () => {
     cy.get('label').contains('Optimal nozzle design').click()
     cy.get('input#nozzleExpansionRatio').type(8)
 
-    cy.get('input#densityRatio').clear().type(0.96)
-    cy.get('input#combustionEfficiencyRatio').clear().type(0.97)
+    cy.get('input#densityRatio').clear()
+    cy.get('input#densityRatio').type(0.96)
 
-    cy.get('div.v-card__title').contains('Advanced settings').parent().contains('Save').click()
+    cy.get('input#combustionEfficiencyRatio').clear()
+    cy.get('input#combustionEfficiencyRatio').type(0.97)
+
+    cy.contains('Advanced settings').parent().contains('Save').click()
 })
 
 Cypress.Commands.add('fillEndBurnerForm', (formValues, units, submit = true) => {
     applyCommonCheck(formValues, units)
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .as('coreSurfaces').eq(0)
         .click()
 
     cy.contains(formValues.propellantId).click()
 
     cy.get('@coreSurfaces').eq(1).click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains('End burner').click()
 
     cy.get('input#throatDiameter')
         .type(formValues.throatDiameter)
+    cy.get('input#throatDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberInnerDiameter')
         .type(formValues.chamberInnerDiameter)
+    cy.get('input#chamberInnerDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#chamberLength')
         .type(formValues.chamberLength)
+    cy.get('input#chamberLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#endBurnerSegmentLength')
         .type(formValues.segmentLength)
+    cy.get('input#endBurnerSegmentLength')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#endBurnerOuterDiameter')
         .type(formValues.outerDiameter)
+    cy.get('input#endBurnerOuterDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#endBurnerHoleDiameter')
         .type(formValues.holeDiameter)
+    cy.get('input#endBurnerHoleDiameter')
         .parent()
         .contains(getLengthUnit(units))
 
     cy.get('input#endBurnerHoleDepth')
         .type(formValues.holeDepth)
+    cy.get('input#endBurnerHoleDepth')
         .parent()
         .contains(getLengthUnit(units))
 
@@ -595,12 +650,12 @@ Cypress.Commands.add('checkPerformanceResults', (expectedResults, units) => {
 
 Cypress.Commands.add("addPropellant", (propellant, unit, closeForm = true) => {
 
-    cy.get('div.v-toolbar__content').contains('Custom propellant')
+    cy.get('.v-dialog--fullscreen > .v-overlay__content > .v-card > .v-card-title').contains('Custom propellant')
 
-    cy.get('.v-select__selections')
+    cy.get('.v-select--single')
         .eq(3)
         .click()
-    cy.get('div.menuable__content__active')
+    cy.get('.v-overlay__content > .v-list')
         .contains(getUnitLabel(unit)).click()
 
     cy.get('input#propellantName')
@@ -609,6 +664,7 @@ Cypress.Commands.add("addPropellant", (propellant, unit, closeForm = true) => {
     if (propellant.cstar) {
         cy.get('input#cstar')
             .type(propellant.cstar)
+        cy.get('input#cstar')
             .parent()
             .contains(getCstarUnit(unit))
     } else {
@@ -632,11 +688,13 @@ Cypress.Commands.add("addPropellant", (propellant, unit, closeForm = true) => {
 
     cy.get('input#density')
         .type(propellant.density)
+    cy.get('input#density')
         .parent()
         .contains(getDensityUnit(unit))
 
     cy.get('input#molarMass')
         .type(propellant.molarMass)
+    cy.get('input#molarMass')
         .parent()
         .contains('kg/kmol')
 
@@ -672,21 +730,22 @@ Cypress.Commands.add('addComplexeBurnRate', (burnRateDataSet, unit) => {
 
         cy.get('input#startPressureInput')
             .type(item.fromPressureIncluded)
-            .parent()
+        cy.get('input#startPressureInput').parent()
             .contains(getPressureUnit(unit))
 
         cy.get('input#endPressureInput')
             .type(item.toPressureExcluded)
+        cy.get('input#endPressureInput')
             .parent()
             .contains(getPressureUnit(unit))
 
         cy.get('input#burnRateCoeffInput')
             .type(item.burnRateCoefficient)
+        cy.get('input#burnRateCoeffInput')
             .parent()
 
-        cy.get('input#pressureExponentInput')
-            .type(item.pressureExponent)
-            .parent()
+        cy.get('input#pressureExponentInput').type(item.pressureExponent)
+        cy.get('input#pressureExponentInput').parent()
 
         cy.get('#saveBurnRateDataBtn')
             .click()
