@@ -1,7 +1,6 @@
 <template>
     <v-container>
         <v-form ref="formJSRM">
-
             <motor-configuration
                 v-model="formValue"
                 @resetValidation="$refs.formJSRM.resetValidation()"
@@ -9,29 +8,33 @@
                 :units="units" ref="motorConfiguration"/>
             <advanced-configuration ref="advanceSettings" v-model="extraConfig" @reset="resetConfig" :units="units"/>
 
-            <div class="text-center" v-if="!disabledButtons">
-                <v-btn class="mr-4" @click="reset">Reset</v-btn>
-                <v-btn class="mr-4" @click="$refs.advanceSettings.show()">
+            <div class="text-center" v-if="!disabledButtons" id="runComputationToolbar">
+                <v-btn class="mr-4" @click="reset" variant="tonal">
+                    Reset
+                </v-btn>
+                <v-btn class="mr-4" @click="$refs.advanceSettings.show()" variant="tonal">
                     <v-icon dark id="btnAdvancedSettings">mdi-cog</v-icon>
                 </v-btn>
-                <v-btn class="mr-4" @click="checkDonor" color="primary" :loading="loading" >Submit</v-btn>
+                <v-btn class="mr-4" @click="checkDonor" color="primary" :loading="loading" variant="tonal">
+                    Submit
+                </v-btn>
             </div>
 
             <v-dialog ref="errorModal" max-width="700px" v-model="showError">
 
                 <v-card>
                     <v-card-title
-                        class="headline grey lighten-2"
+                        class="text-h5 bg-grey-lighten-2"
                         primary-title>
                         Computation failed
                     </v-card-title>
 
                     <v-card-text class="mt-5">
-                        <v-textarea readonly outline
+                        <v-textarea readonly variant="outlined"
                                     v-show="errorDetail == null"
                                     v-model="errorMessage"/>
 
-                        <v-textarea readonly outline
+                        <v-textarea readonly variant="outlined"
                                     :label="errorMessage"
                                     v-show="errorDetail !== null"
                                     v-model="errorDetail"/>
@@ -55,13 +58,13 @@
                 </v-card>
             </v-dialog>
         </v-form>
-        <v-overlay :value="loading">
-            <v-col>
-                <div class="text-center">
-                    <v-progress-circular indeterminate size="64" width="6"></v-progress-circular>
-                </div>
-                <div class="text-center mt-30">Computation in progress ...</div>
-            </v-col>
+        <v-overlay :model-value="loading" class="align-center justify-center">
+            <v-row>
+                <v-col align-self="center" align="center">
+                    <v-progress-circular class="align-center" indeterminate size="64" color="purple" width="6"></v-progress-circular>
+                    <div class="text-center mt-30"><h3 class="text-white">Computation in progress ...</h3></div>
+                </v-col>
+            </v-row>
         </v-overlay>
         <donate ref="donationPopup" :check-mode="true" @closeDonation="runComputation"></donate>
     </v-container>
@@ -118,8 +121,13 @@ export default {
         getDateTimeAsString() {
             return new Date().toLocaleString()
         },
-        validateForm() {
-            return this.$refs.formJSRM.validate()
+        validateForm(functionToExecute = () => console.error('NOT implemented check validateForm()')) {
+            this.$refs.formJSRM.validate()
+                .then(result => {
+                    if (result.valid) {
+                        functionToExecute()
+                    }
+                })
         },
         checkGrainAndGetURL() {
             let url = '/compute'
@@ -153,19 +161,24 @@ export default {
         },
         checkDonor() {
             const checkGrainAndGetURL = this.checkGrainAndGetURL()
-            if (this.$refs.formJSRM.validate() && checkGrainAndGetURL.grainCheck) {
-                this.$refs.donationPopup.check()
+            if (checkGrainAndGetURL.grainCheck) {
+                this.$refs.formJSRM.validate()
+                    .then((result) => {
+                        if(result.valid) {
+                            this.$refs.donationPopup.check()
+                        }
+                    })
             }
         },
         ...mapMutations('computation', ['setCurrentComputation', 'setPreviousComputation', 'saveCurrentMotor']),
         ...mapGetters('computation', ['currentComputation', 'isPreviousMotorFlagAsReference']),
         ...mapGetters('authentication', ['isDonator']),
-        runComputation() {
+        runComputationExecutor() {
             const component = this
             let request
             const checkGrainAndGetURL = this.checkGrainAndGetURL()
-            request = this.buildRequest()
-            if (this.$refs.formJSRM.validate() && checkGrainAndGetURL.grainCheck) {
+            if (checkGrainAndGetURL.grainCheck) {
+                request = this.buildRequest()
                 this.loading = true
                 Axios.post(checkGrainAndGetURL.url, request)
                     .then((response) => {
@@ -197,6 +210,13 @@ export default {
                     })
             }
         },
+        runComputation() {
+            this.$refs.formJSRM.validate().then(result => {
+                if (result.valid) {
+                    this.runComputationExecutor()
+                }
+            })
+        },
         checkMotorDimensions() {
             if (this.formValue.chamberLength < this.formValue.grainConfig.segmentLength * this.formValue.grainConfig.numberOfSegment) {
                 this.errorDetail = `The 'combustion chamber length' should be >= 'grain segment length' times 'number of segment'. Otherwise your grain configuration will not fit into your motor. Increase your combustion chamber length and/or decrease : grain segment length, number of segment.`
@@ -207,35 +227,27 @@ export default {
         },
 
         buildExport() {
-            if (this.$refs.formJSRM.validate()) {
-                let request = Object.assign({ }, this.formValue)
-                request.extraConfig = Object.assign({}, this.extraConfig)
-                request.measureUnit = this.units.type
-                delete request.measureUnit
-                return request
-            } else {
-                return null
-            }
+            let request = Object.assign({ }, this.formValue)
+            request.extraConfig = Object.assign({}, this.extraConfig)
+            request.measureUnit = this.units.type
+            delete request.measureUnit
+            return request
         },
         buildRequest() {
-            if (this.$refs.formJSRM.validate()) {
-                let request = Object.assign({ }, this.formValue)
-                request.extraConfig = Object.assign({}, this.extraConfig)
-                request.measureUnit = this.units.type
+            let request = Object.assign({ }, this.formValue)
+            request.extraConfig = Object.assign({}, this.extraConfig)
+            request.measureUnit = this.units.type
 
-                // save cuurent config to restore it later if necessary
-                this.saveCurrentMotor(Object.assign({ }, request))
+            // save cuurent config to restore it later if necessary
+            this.saveCurrentMotor(Object.assign({ }, request))
 
-                delete request.grainType
-                request = Object.assign(request, request.grainConfig)
-                delete request.grainConfig
+            delete request.grainType
+            request = Object.assign(request, request.grainConfig)
+            delete request.grainConfig
 
-                request.grainType = this.getGrainType()
+            request.grainType = this.getGrainType()
 
-                return request
-            } else {
-                return null
-            }
+            return request
         },
         loadForm(formData = {}, extraConfig = this.getDefaultAdvanceConfig()) {
             this.extraConfig = extraConfig
