@@ -1,27 +1,24 @@
 <template>
-    <v-container>
-        <v-layout column align-center>
-            <v-flex xs10 sm6>
-                <v-icon size="80">mdi-rocket</v-icon>
-            </v-flex>
-            <v-flex xs10 sm6 >
-            <h1>
-                Sign in to METEOR
-            </h1>
-            </v-flex>
-            <v-flex grow>
-                <v-alert
-                    v-if="showError"
-                    border="top"
-                    colored-border
-                    type="error"
-                    elevation="2"
-                    max-width="400"
-                >
-                    Authentication failed
-                </v-alert>
+    <v-container fluid align="center">
+            <v-col grow xs="10" sm="6" lg="3">
                 <v-card>
+                    <v-card-title>
+                        <v-icon size="80" color="purple">mdi-rocket-launch</v-icon>
+                        <h2>
+                            Sign in to METEOR
+                        </h2>
+                    </v-card-title>
                     <v-card-text class="mt-5">
+                        <v-alert
+                            v-if="showError"
+                            border="start"
+                            border-color="top"
+                            variant="outlined"
+                            type="error"
+                            class="mb-5"
+                        >
+                            Authentication failed
+                        </v-alert>
                         <v-form
                             ref="form"
                             v-model="valid"
@@ -38,13 +35,12 @@
                                 id="signinPassword"
                                 v-model="password"
                                 label="Password"
-                                :append-icon="showPassword ? 'mdi-eye' : 'mdi-eye-off'"
+                                :append-inner-icon="showPassword ? 'mdi-eye' : 'mdi-eye-off'"
                                 :type="showPassword ? 'text' : 'password'"
                                 :rules="passwordRules"
-                                @click:append="showPassword = !showPassword"
+                                @click:append-inner="showPassword = !showPassword"
                                 required
                             ></v-text-field>
-                            <router-link :to="'/lost-password'">Forgot password?</router-link>
                             <v-btn width="100%"
                                    style="margin-top: 20px"
                                 :disabled="!valid"
@@ -58,16 +54,12 @@
                         </v-form>
                     </v-card-text>
                     <v-card-actions>
+                        <router-link to="/signup">Create an account</router-link>
                         <v-spacer></v-spacer>
+                        <router-link :to="'/lost-password'">Forgot password?</router-link>
                     </v-card-actions>
                 </v-card>
-                <v-card>
-                    <v-card-text class="text--primary">
-                        New to METEOR? <v-btn color="blue" :to="'/signup'" text>Create an account.</v-btn>
-                    </v-card-text>
-                </v-card>
-            </v-flex>
-        </v-layout>
+            </v-col>
     </v-container>
 </template>
 
@@ -98,25 +90,27 @@ export default {
             }
         },
         signin() {
-            if (this.$refs.form && this.$refs.form.validate()) {
-                const me = this
-                me.showError = false
-                me.loading = true
-                Axios.post('/auth/signin', {
-                    username: this.email,
-                    password: this.password
+            if (this.$refs.form) {
+                this.$refs.form.validate().then(() => {
+                    const me = this
+                    me.showError = false
+                    me.loading = true
+                    Axios.post('/auth/signin', {
+                        username: this.email,
+                        password: this.password
+                    })
+                        .then(function(response) {
+                            me.saveToken(response.data)
+                            me.$router.push({ path: '/motorDesign' })
+                            me.loading = false
+                        })
+                        .catch((error) => {
+                            me.password = ''
+                            me.showError = true
+                            me.loading = false
+                            console.error(error)
+                        })
                 })
-                    .then(function(response) {
-                        me.saveToken(response.data)
-                        me.$router.push({ path: '/motorDesign' })
-                        me.loading = false
-                    })
-                    .catch((error) => {
-                        me.password = ''
-                        me.showError = true
-                        me.loading = false
-                        console.error(error)
-                    })
             }
         },
         ...mapActions('authentication', ['saveToken'])
