@@ -1,72 +1,68 @@
 <template>
-    <v-container>
-        <v-layout column align-center>
-            <v-flex xs10 sm6>
-                <v-icon size="80">mdi-rocket</v-icon>
-            </v-flex>
-            <v-flex xs10 sm6 >
-                <h1>
-                    Create an account
-                </h1>
-            </v-flex>
-            <v-flex grow>
-                <v-alert
-                    v-if="showMessage"
-                    border="top"
-                    colored-border
-                    :type="messageType"
-                    elevation="2"
-                    max-width="400"
-                >
-                    {{ message }}
-                </v-alert>
-                <v-card v-if="!emailSent">
-                    <v-card-text class="mt-5">
-                        <v-form
-                            ref="form"
-                            v-model="valid"
+    <v-container fluid align="center">
+        <v-col grow xs="10" sm="6" lg="3">
+            <v-alert
+                v-show="showMessage"
+                border="start"
+                border-color="top"
+                :type="messageType"
+                variant="outlined"
+                class="mb-5"
+            >
+                {{ message }}
+            </v-alert>
+            <v-card v-if="!emailSent">
+                <v-card-title>
+                    <v-icon size="80" color="purple">mdi-rocket-launch</v-icon>
+                    <h2>
+                        Create an account
+                    </h2>
+                </v-card-title>
+                <v-card-text class="mt-5">
+                    <v-form
+                        ref="form"
+                        v-model="valid"
+                    >
+                        <v-text-field
+                            id="signupEmail"
+                            v-model="email"
+                            label="E-mail"
+                            :rules="emailRules"
+                            required
+                        ></v-text-field>
+                        <v-text-field
+                            id="signupPassword"
+                            v-model="password"
+                            label="Password"
+                            :append-inner-icon="showPassword ? 'mdi-eye' : 'mdi-eye-off'"
+                            :type="showPassword ? 'text' : 'password'"
+                            :rules="passwordRules"
+                            @click:append-inner="showPassword = !showPassword"
+                            required
+                        ></v-text-field>
+                        <v-text-field
+                            id="signupPasswordConfirm"
+                            v-model="passwordConfirm"
+                            label="Confirm password"
+                            :append-inner-icon="showPassword2 ? 'mdi-eye' : 'mdi-eye-off'"
+                            :type="showPassword2 ? 'text' : 'password'"
+                            :rules="confirmPasswordRule"
+                            @click:append-inner="showPassword2 = !showPassword2"
+                            required
+                        ></v-text-field>
+                        <v-btn
+                            width="100%"
+                            :disabled="!valid"
+                            color="success"
+                            @click="signup"
+                            :loading="loading"
                         >
-                            <v-text-field
-                                id="signupEmail"
-                                v-model="email"
-                                label="E-mail"
-                                :rules="emailRules"
-                                required
-                            ></v-text-field>
-                            <v-text-field
-                                id="signupPassword"
-                                v-model="password"
-                                label="Password"
-                                :append-icon="showPassword ? 'mdi-eye' : 'mdi-eye-off'"
-                                :type="showPassword ? 'text' : 'password'"
-                                :rules="passwordRules"
-                                @click:append="showPassword = !showPassword"
-                                required
-                            ></v-text-field>
-                            <v-text-field
-                                id="signupPasswordConfirm"
-                                v-model="passwordConfirm"
-                                label="Confirm password"
-                                :append-icon="showPassword2 ? 'mdi-eye' : 'mdi-eye-off'"
-                                :type="showPassword2 ? 'text' : 'password'"
-                                :rules="confirmPasswordRule"
-                                @click:append="showPassword2 = !showPassword2"
-                                required
-                            ></v-text-field>
-                            <v-btn
-                                width="100%"
-                                :disabled="!valid"
-                                color="success"
-                                @click="signup"
-                                :loading="loading"
-                            >
-                                Create account
-                            </v-btn>
-                        </v-form>
-                    </v-card-text>
-                </v-card>
-            </v-flex>
-        </v-layout>
+                            Create account
+                        </v-btn>
+                    </v-form>
+                </v-card-text>
+            </v-card>
+        </v-col>
     </v-container>
 
 </template>
@@ -99,27 +95,29 @@ export default {
             return value => value === password || 'Password don\'t match'
         },
         signup() {
-            if (this.$refs.form.validate() && this.password === this.passwordConfirm) {
-                const me = this
-                me.loading = true
-                me.clearToken()
-                Axios.post('/auth/signup', {
-                    email: this.email,
-                    password: this.password
-                })
-                    .then(function() {
-                        me.message = 'An activation link has been sent to your address. Also please check the Spam folder in your mailbox.'
-                        me.messageType = 'success'
-                        me.showMessage = true
-                        me.emailSent = true
-                        me.loading = false
-                    })
-                    .catch(function(error) {
-                        me.message = error.response.data.message
-                        me.messageType = 'error'
-                        me.showMessage = true
-                        me.loading = false
-                        console.error(error)
+            if (this.password === this.passwordConfirm) {
+                this.$refs.form.validate()
+                    .then(() => {
+                        this.loading = true
+                        this.clearToken()
+                        Axios.post('/auth/signup', {
+                            email: this.email,
+                            password: this.password
+                        })
+                            .then(() => {
+                                this.message = 'An activation link has been sent to your address. Also please check the Spam folder in your mailbox.'
+                                this.messageType = 'success'
+                                this.showMessage = true
+                                this.emailSent = true
+                                this.loading = false
+                            })
+                            .catch((error) => {
+                                this.message = error.response.data.message
+                                this.messageType = 'error'
+                                this.showMessage = true
+                                this.loading = false
+                                console.error(error)
+                            })
                     })
             }
         },
