@@ -1,20 +1,22 @@
 <template>
     <v-dialog v-model="sheet" width="500" persistent>
-        <template v-slot:activator="{ on }" v-show="!checkMode">
+        <template v-slot:activator="{ props }">
             <v-btn
+                color="white"
                 id="btnDonate"
-                color="purple"
                 v-show="!checkMode"
-                 dark
-                 v-on="on"
+                v-bind="props"
+                variant="text"
+                style="background-color: #9c27b0"
+                width="140px"
             >
-                <v-icon left size="25">mdi-handshake</v-icon>
+                <v-icon start size="25">mdi-handshake</v-icon>
                 Donate
             </v-btn>
         </template>
         <v-card>
             <v-card-title
-                class="headline purple"
+                class="text-h5 bg-purple"
                 primary-title
                 style="color: white"
             >
@@ -22,7 +24,7 @@
             </v-card-title>
             <v-card-text>
                 <v-row justify="center" align="center">
-                    <v-flex shrink>
+                    <v-col shrink>
                         <v-col>
                             <div style="text-align: center; padding: 20px; margin: 20px 40px 0 40px; border: 1px solid dimgray; border-radius: 5px;">
                                 <h2 style="margin-bottom: 10px;">Last 30 days donations</h2>
@@ -33,11 +35,10 @@
                                 <ul style="margin-bottom: 5px">
                                     <li>
                                         <b>do not see this popup anymore</b>
-                                        <v-tooltip bottom>
-                                            <template v-slot:activator="{ on, attrs }">
+                                        <v-tooltip location="bottom">
+                                            <template v-slot:activator="{props}">
                                                 <v-icon
-                                                    v-bind="attrs"
-                                                    v-on="on"
+                                                    v-bind="props"
                                                 >
                                                     mdi-information-variant
                                                 </v-icon>
@@ -56,9 +57,9 @@
                                 </p>
                             </div>
                         </v-col>
-                    </v-flex>
+                    </v-col>
                 </v-row>
-                <v-alert type="info" v-model="donateClicked" outlined>
+                <v-alert type="info" v-model="donateClicked" variant="outlined">
                     If you have made a donation, you must log out and log back in to be marked as donator.
                 </v-alert>
             </v-card-text>
@@ -66,7 +67,7 @@
             <v-card-actions>
                 <v-spacer></v-spacer>
                 <v-btn
-                    outlined
+                    variant="outlined"
                     v-if="!donateClicked"
                     @click="sheet = false"
                     :loading="!closable"
@@ -84,12 +85,12 @@
                     v-else
                     id="btnDonateNow"
                     color="purple"
-                    dark
+                    theme="dark"
                     @click="onDonateClick()"
                     target="_blank"
                     href="https://pages.donately.com/meteor/campaign/meteor/donate"
                 >
-                    <v-icon left size="25"  v-on="on">mdi-handshake</v-icon>
+                    <v-icon start size="25"  v-bind="props">mdi-handshake</v-icon>
                     Donate now
                 </v-btn>
             </v-card-actions>
