@@ -3,13 +3,13 @@
         <v-dialog scrollable v-model="dialog" persistent max-width="350px">
             <v-card>
                 <v-card-title
-                    class="headline grey lighten-2"
+                    class="text-h5 bg-grey-lighten-2"
                     primary-title>
                     Burn data editor
                 </v-card-title>
                 <v-card-text>
                     <v-layout row wrap>
-                        <v-flex>
+                        <v-col>
                                 <v-form ref="burnRateEditorForm" class="mr-5 ml-5">
                                     <v-text-field id="startPressureInput" label="From pressure (included):" v-model="burnData.fromPressureIncluded" :suffix="units.pressureUnit" :rules="numberRule"/>
                                     <v-text-field id="endPressureInput" label="To pressure (excluded):" v-model="burnData.toPressureExcluded" :suffix="units.pressureUnit" :rules="numberRule" :error-messages="endPressureErrorMessage"/>
@@ -17,13 +17,13 @@
                                     <v-text-field id="pressureExponentInput" label="Pressure exponent:" v-model="burnData.pressureExponent" :hint="hintBurnRate" persistent-hint :rules="numberRule"/>
                                 </v-form>
 
-                        </v-flex>
+                        </v-col>
                     </v-layout>
                 </v-card-text>
                 <v-card-actions>
                     <v-spacer></v-spacer>
                     <v-btn @click="cancel">Cancel</v-btn>
-                    <v-btn color="primary" @click="save" id="saveBurnRateDataBtn">Save</v-btn>
+                    <v-btn color="primary" @click="isFormValid" id="saveBurnRateDataBtn">Save</v-btn>
                 </v-card-actions>
             </v-card>
         </v-dialog>
@@ -32,13 +32,15 @@
 
 <script>
 import { numberRule } from '../../modules/formValidationRules'
-import Vue from 'vue'
+import { nextTick } from 'vue'
+import { object } from '@amcharts/amcharts4/core'
 export default {
     name: 'BurnDataEditor',
     data() {
         return {
             dialog: false,
             burnData: {},
+            originalBurnData: {},
             hintBurnRate: 'Value is different between SI and Imperial',
             createMode: false,
             numberRule: numberRule()
@@ -55,23 +57,29 @@ export default {
             this.createMode = true
             this.burnData = {}
             this.show()
-            Vue.nextTick(() => this.$refs.burnRateEditorForm.reset())
+            nextTick(() => this.$refs.burnRateEditorForm.reset())
         },
         edit(item) {
             this.createMode = false
-            this.burnData = item
+            this.originalBurnData = item
+            this.burnData = object.clone(item)
             this.show()
         },
         save() {
-            if (this.isFormValid()) {
-                if (this.createMode) {
-                    this.$emit('created', this.burnData)
-                }
-                this.dialog = false
+            if (this.createMode) {
+                this.$emit('created', this.burnData)
+            } else {
+                Object.assign(this.originalBurnData, this.burnData)
             }
+            this.dialog = false
         },
         isFormValid() {
-            return this.$refs.burnRateEditorForm ? this.$refs.burnRateEditorForm.validate() : true
+            this.$refs.burnRateEditorForm.validate()
+                .then(result => {
+                    if (result.valid) {
+                        this.save()
+                    }
+                })
         },
         cancel() {
             this.burnData = {}
