@@ -1,67 +1,67 @@
 <template>
     <div v-if="!!performance && !redraw">
-        <v-layout d-flex wrap>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="motor-class" :suffix="performance.classPercentage+'%'"
-                          label="Class" v-model="performance.motorDescription" :messages="getPreviousValue('motorDescription')">>
+        <v-row d-flex wrap>
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="motor-class" :suffix="performance.classPercentage+'%'"
+                          label="Class" v-model="performance.motorDescription" :messages="getPreviousValue('motorDescription')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors" >
                     <v-icon :color="compareMotors.class.cssColor">{{ compareMotors.class.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="thrust-time" label="Thrust time" suffix="s"
-                          v-model="performance.thrustTime" :messages="getPreviousValue('thrustTime')">
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="thrust-time" label="Thrust time" suffix="s"
+                          v-model="performance.thrustTime" :messages="getPreviousValue('thrustTime')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.thrustTime.cssColor">{{ compareMotors.thrustTime.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="max-thrust" label="Max thrust" suffix="N"
-                          v-model="performance.maxThrust" :messages="getPreviousValue('maxThrust')">>
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="max-thrust" label="Max thrust" suffix="N"
+                          v-model="performance.maxThrust" :messages="getPreviousValue('maxThrust')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.maxThrust.cssColor">{{ compareMotors.maxThrust.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="total-impulse" label="Total impulse" suffix="Ns"
-                          v-model="performance.totalImpulse" :messages="getPreviousValue('totalImpulse')">>
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="total-impulse" label="Total impulse" suffix="Ns"
+                          v-model="performance.totalImpulse" :messages="getPreviousValue('totalImpulse')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.totalImpulse.cssColor">{{ compareMotors.totalImpulse.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="specific-impulse" label="Specific impulse" suffix="s"
-                          v-model="performance.specificImpulse" :messages="getPreviousValue('specificImpulse')">>
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="specific-impulse" label="Specific impulse" suffix="s"
+                          v-model="performance.specificImpulse" :messages="getPreviousValue('specificImpulse')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.specificImpulse.cssColor">{{ compareMotors.specificImpulse.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="max-pressure" label="Max pressure" :suffix="units.resultPressureUnit"
-                          v-model="performance.maxPressure" :messages="getPreviousValue('maxPressure')">
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="max-pressure" label="Max pressure" :suffix="units.resultPressureUnit"
+                          v-model="performance.maxPressure" :messages="getPreviousValue('maxPressure')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.maxPressure.cssColor">{{ compareMotors.maxPressure.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="average-pressure" label="Average pressure"
-                          :suffix="units.resultPressureUnit" v-model="performance.averagePressure" :messages="getPreviousValue('averagePressure')">>
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="average-pressure" label="Average pressure"
+                          :suffix="units.resultPressureUnit" v-model="performance.averagePressure" :messages="getPreviousValue('averagePressure')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.averagePressure.cssColor">{{ compareMotors.averagePressure.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="nozzle-exit-speed" label="Nozzle exit speed" prefix="Mach"
-                          v-model="performance.exitSpeedInitial" :messages="getPreviousValue('exitSpeedInitial')">
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="nozzle-exit-speed" label="Nozzle exit speed" prefix="Mach"
+                          v-model="performance.exitSpeedInitial" :messages="getPreviousValue('exitSpeedInitial')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.exitSpeedInitial.cssColor">{{ compareMotors.exitSpeedInitial.icon }}</v-icon>
                 </template>
             </v-text-field>
-            <v-text-field :hide-details="!compareWithPrevious" outlined class="resultElement" readonly id="grain-mass" label="Initial grain mass"
-                          :suffix="units.grainMassUnit" v-model="performance.grainMass" :messages="getPreviousValue('grainMass')">>
+            <v-text-field :hide-details="!compareWithPrevious" variant="outlined" class="resultElement" readonly id="grain-mass" label="Initial grain mass"
+                          :suffix="units.grainMassUnit" v-model="performance.grainMass" :messages="getPreviousValue('grainMass')" density="compact">
                 <template v-slot:prepend-inner v-if="!!compareMotors">
                     <v-icon :color="compareMotors.grainMass.cssColor">{{ compareMotors.grainMass.icon }}</v-icon>
                 </template>
             </v-text-field>
-        </v-layout>
-        <v-flex >
+        </v-row>
+        <v-col >
             <p class="label-nozzle" v-bind:style="{color: this.portThroatWarningColor}" v-if="showPortThroatAreaWarning">
                 {{`Your port-to-throat ratio is ${currentComputation.performanceResult.portToThroatArea}, it can be a problem` }}
-                <v-tooltip bottom>
-                    <template v-slot:activator="{ on }">
-                        <v-icon v-on="on">mdi-information-outline</v-icon>
+                <v-tooltip location="bottom">
+                    <template v-slot:activator="{ props }">
+                        <v-icon v-bind="props">mdi-information-outline</v-icon>
                     </template>
                     <p class="label-nozzle tooltip-performance">
                         Enlarging the core is a good way to solve this problem. Except for extreme L/D ratio motors, no part of the core should be less than the diameter of the nozzle throat.
@@ -70,9 +70,9 @@
             </p>
             <p class="label-nozzle" style="color: red;" v-if="currentComputation.performanceResult.lowKNCorrection">
                 Low KN.
-                <v-tooltip bottom>
-                    <template v-slot:activator="{ on }">
-                        <v-icon v-on="on">mdi-information-outline</v-icon>
+                <v-tooltip location="bottom">
+                    <template v-slot:activator="{ props }">
+                        <v-icon v-bind="props">mdi-information-outline</v-icon>
                     </template>
                     <p class="label-nozzle tooltip-performance">
                         KN is ratio between burn area and throat area. With low KN the result is approximate and the motor may be subject of chuffing. To increase it you can use a larger grain core diameter and/or decrease the throat diameter.
@@ -80,7 +80,7 @@
                 </v-tooltip>
             </p>
             <p class="label-nozzle" style="color: #2196f3;" v-if="currentComputation.performanceResult.optimalDesign">Optimally designed nozzle with an expansion ratio of {{currentComputation.performanceResult.optimalNozzleExpansionRatio}}</p>
-        </v-flex>
+        </v-col>
     </div>
 
 </template>
@@ -88,7 +88,7 @@
 <script>
 import { mapGetters } from 'vuex'
 import { comparePerformanceResults } from '@/modules/computationUtils.mjs'
-import Vue from 'vue'
+import { nextTick } from 'vue'
 
 export default {
     name: 'PerformanceInfo',
@@ -124,7 +124,7 @@ export default {
     watch: {
         compareMotors() {
             this.redraw = true
-            Vue.nextTick(() => { this.redraw = false })
+            nextTick(() => { this.redraw = false })
         },
         currentComputation(newValue) {
             this.performance = newValue.performanceResult
