@@ -24,7 +24,7 @@ const actions = {
                 showError('Failed to retrieve propellant list')
             })
     },
-    deletePropellant({ dispatch, commit }, { propellant, showError = (message) => console.error(message), successCallback = () => {} }) {
+    deletePropellant({ dispatch }, { propellant, showError = (message) => console.error(message), successCallback = () => {} }) {
         Axios.delete(`/propellants/${propellant.id}`)
             .then(successCallback)
             .catch(function(error) {
