@@ -24,7 +24,7 @@ const actions = {
                 showError('Failed to retrieve motor list')
             })
     },
-    deleteMotor({ dispatch, commit }, { motor, showError = (message) => console.error(message) }) {
+    deleteMotor({ dispatch }, { motor, showError = (message) => console.error(message) }) {
         Axios.delete(`/motors/${motor.id}`)
             .catch(function(error) {
                 console.error(error)
